Module jenkins_api::action

Types to parse the actions that triggered a Build

Modules

causes
Types to parse the causes of a Build

git
Types related to git

maven
Types related to maven

parameters
Types to parse the parameters of a Build

pipeline
Nodes found in a pipeline

Structs

CauseAction
An action listing causes

CommonAction
A node of a pipeline

EnvActionImpl
An action from pipelines

FlowGraphAction
An action from pipelines

GitBuildData
An action describing a Git change

GitTagAction
An action for a git tag

MavenAggregatedArtifactRecord
An action with maven artifacts

MavenArtifactRecord
An action with maven artifacts

ParametersAction
An action holding parameters

PipelineApproverAction
An action marking an approval on a pipeline

RepoTagAction
An action for a repo tag

SurefireAggregatedReport
An action with a surefire test report aggregated from other reports

SurefireReport
An action with a surefire test report

TimeInQueueAction
An action on time in queue

Traits

Action
Trait implemented by specialization of Action