Module jenkins_api::job

Jenkins Jobs

Modules

builder
Helper to build a job

Structs

BuildFlowJob
A build flow job

CommonJob
A Jenkins Job

ExternalJob
An external job

Folder
A folder

FreeStyleProject
A free style project

HealthReport
Health Report of a Job

JobName
Helper type to act on a job

MatrixConfiguration
A matrix configuration

MatrixProject
A matrix project

MavenModule
A maven module

MavenModuleSet
A maven project

MultiJobProject
A MultiJob Project

ShortJob
Short Job that is used in lists and links from other structs

WorkflowJob
A pipeline project

WorkflowMultiBranchProject
A pipeline project

Enums

BallColor
Ball Color corresponding to a BuildStatus

Traits

BuildableJob
Common trait for jobs that can be build

Job
Trait implemented by specializations of Job and providing common methods

SCMPollable
Common trait for jobs that can poll a SCM