Trait jenkins_api::job::Job

pub trait Job {
    fn url(&self) -> &str;

    fn name(&self) -> &str;

    fn enable(&self, jenkins_client: &Jenkins) -> Result<()> { ... }

    fn disable(&self, jenkins_client: &Jenkins) -> Result<()> { ... }

    fn add_to_view<'a, V>(
        &self, 
        jenkins_client: &Jenkins, 
        view_name: V
    ) -> Result<()>
    where
        V: Into<ViewName<'a>>,
    { ... }

    fn remove_from_view<'a, V>(
        &self, 
        jenkins_client: &Jenkins, 
        view_name: V
    ) -> Result<()>
    where
        V: Into<ViewName<'a>>,
    { ... }

    fn get_config_xml(&self, jenkins_client: &Jenkins) -> Result<String> { ... }
}

Trait implemented by specializations of Job and providing common methods

Required methods

fn url(&self) -> &str

get the url of a Job

fn name(&self) -> &str

Get the name of the project

Loading content...
Provided methods

fn enable(&self, jenkins_client: &Jenkins) -> Result<()>

Enable a Job. It may need to be refreshed as it may have been updated

fn disable(&self, jenkins_client: &Jenkins) -> Result<()>

Disable a Job. It may need to be refreshed as it may have been updated

fn add_to_view<'a, V>(
&self,
jenkins_client: &Jenkins,
view_name: V
) -> Result<()> where
V: Into<ViewName<'a>>,

Add this job to the view view_name

fn remove_from_view<'a, V>(
&self,
jenkins_client: &Jenkins,
view_name: V
) -> Result<()> where
V: Into<ViewName<'a>>,

Remove this job from the view view_name

fn get_config_xml(&self, jenkins_client: &Jenkins) -> Result<String>

Get the config.xml file for this job

Loading content...