Struct jenkins_api::Jenkins

pub struct Jenkins { /* fields omitted */ }

Client struct with the methods to query Jenkins

Implementations

impl Jenkins

pub fn get_object_as<Q, T>(&self, object: Path, parameters: Q) -> Result<T> where
Q: Into<Option<AdvancedQuery>>,
T: Deserialize<'de>,

Get a Path from Jenkins, specifying the depth or tree parameters

Example

#[derive(Deserialize)]
#[serde(rename_all = "camelCase")]
struct LastBuild {
    number: u32,
    duration: u32,
    result: String,
}
#[derive(Deserialize)]
#[serde(rename_all = "camelCase")]
struct LastBuildOfJob {
    display_name: String,
    last_build: LastBuild,
}

let _: LastBuildOfJob = jenkins.get_object_as(
    jenkins_api::client::Path::Job {
        name: "job name",
        configuration: None,
    },
    jenkins_api::client::TreeBuilder::new()
        .with_field("displayName")
        .with_field(
            jenkins_api::client::TreeBuilder::object("lastBuild")
                .with_subfield("number")
                .with_subfield("duration")
                .with_subfield("result"),
        )
        .build(),
)?;

impl Jenkins

pub fn get_build<'a, J, B>(
&self,
job_name: J,
build_number: B
) -> Result<CommonBuild> where
J: Into<JobName<'a>>,
B: Into<BuildNumber>,

Get a build from a job_name and build_number

impl Jenkins

pub fn get_home(&self) -> Result<Home>

Get Jenkins Home

impl Jenkins

pub fn get_job<'a, J>(&self, job_name: J) -> Result<CommonJob> where
J: Into<JobName<'a>>,

Get a Job from it's job_name

pub fn build_job<'a, J>(&self, job_name: J) -> Result<ShortQueueItem> where
J: Into<JobName<'a>>,

Build a Job from it's job_name

pub fn job_builder<'a, 'b, 'c, 'd>(
&'b self,
job_name: &'a str
) -> Result<JobBuilder<'a, 'b, 'c, 'd>>

Create a JobBuilder to setup a build of a Job from it's job_name

pub fn poll_scm_job<'a, J>(&self, job_name: J) -> Result<()> where
J: Into<JobName<'a>>,

Poll SCM of a Job from it's job_name

impl Jenkins

pub fn get_nodes(&self) -> Result<ComputerSet>

Get a ComputerSet

pub fn get_node<'a, C>(&self, computer_name: C) -> Result<CommonComputer> where
C: Into<ComputerName<'a>>,

Get a Computer

pub fn get_master_node(&self) -> Result<MasterComputer>

Get the master Computer

impl Jenkins

pub fn get_queue(&self) -> Result<Queue>

Get the Jenkins items queue

pub fn get_queue_item(&self, id: i32) -> Result<QueueItem>

Get a queue item from it's ID

impl Jenkins

pub fn get_view<'a, V>(&self, view_name: V) -> Result<CommonView> where
V: Into<ViewName<'a>>,

Get a View

pub fn add_job_to_view<'a, 'b, V, J>(
&self,
view_name: V,
job_name: J
) -> Result<()> where
V: Into<ViewName<'a>>,
J: Into<JobName<'a>>,

Add the job job_name to the view view_name

pub fn remove_job_from_view<'a, 'b, V, J>(
&self,
view_name: V,
job_name: J
) -> Result<()> where
V: Into<ViewName<'a>>,
J: Into<JobName<'a>>,

Remove the job job_name from the view view_name