Module jenkins_api::build

Jenkins Builds

Structs

Artifact
A file archived by a Build

BuildFlowRun
A Build from a BuildFlowJob

CommonBuild
A Jenkins Build

FreeStyleBuild
A Build from a FreeStyleProject

MatrixBuild
A Build from a MatrixProject

MatrixRun
A Build from a MatrixConfiguration

MavenBuild
A Build of a MavenModule

MavenModuleSetBuild
A Build of a MavenModuleSet

MultiJobBuild
A Build from a MultiJobProject

ShortBuild
Short Build that is used in lists and links from other structs

WorkflowRun
A Build from a WorkflowJob

Enums

BuildNumber
Helper type to act on a build

BuildStatus
Status of a build

Traits

Build
Trait implemented by specializations of Build and providing common methods